LSM9DS1 is testable! Results are solid!
-Tsp

+++ LSM9DS1/lsm9ds1.h
@@ -12,12 +12,12 @@ void init_mag();
 void calibrate_mag();
 
 /* General IMU Functions */
-void read_device_bytes(unsigned char addr, unsigned char sub_addr, int* dest);
+void read_device_bytes(unsigned char addr, unsigned char sub_addr, short* dest);
 unsigned char imu_read_byte(unsigned char addr, unsigned char sub_addr);
 void imu_write_byte(unsigned char addr, unsigned char sub_addr, unsigned char val);
 void calibrate_IMU();
 char init_IMU();
-int* IMU_pull_samples();
+float* IMU_pull_samples();
 void read_memory();
 
 #endif  //LSM9DS1_H
